+++ src/d_iwad.c
@@ -187,6 +187,21 @@ static const char *steam_install_subdirs[] =
     "steamapps\\common\\ultimate doom\\base"
 };
 
+// Location where Bethesda.net Launcher is installed
+static registryvalue_t bethesdanet_install_location =
+{
+    HKEY_CURRENT_USER,
+    "Software\\Bethesda Softworks\\Bethesda.net",
+    "installLocation"
+};
+
+// Subdirs of the Bethesda.net Launcher install directory where IWADs are found
+static const char *bethesdanet_install_subdirs[] =
+{
+    "games\\DOOM II\\base",
+    "games\\Ultimate DOOM\\base"
+};
+
 static char *GetRegistryString(registryvalue_t *reg_val)
 {
     HKEY    key;
@@ -278,6 +293,25 @@ static void CheckSteamEdition(void)
     for (size_t i = 0; i < arrlen(steam_install_subdirs); i++)
     {
         char    *path = M_StringJoin(install_path, DIR_SEPARATOR_S, steam_install_subdirs[i], NULL);
+
+        AddIWADDir(path);
+        free(path);
+    }
+
+    free(install_path);
+}
+
+// Check for DOOM downloaded via Bethesda.net Launcher
+static void CheckBethesdaNetEdition(void)
+{
+    char *install_path = GetRegistryString(&bethesdanet_install_location);
+
+    if (!install_path)
+        return;
+
+    for (size_t i = 0; i < arrlen(bethesdanet_install_subdirs); i++)
+    {
+        char *path = M_StringJoin(install_path, DIR_SEPARATOR_S, bethesdanet_install_subdirs[i], NULL);
 
         AddIWADDir(path);
         free(path);
@@ -403,6 +437,7 @@ static void BuildIWADDirList(void)
     CheckUninstallStrings();
     CheckInstallRootPaths();
     CheckSteamEdition();
+    CheckBethesdaNetEdition();
     CheckDOSDefaults();
 #endif